chore(charts): begin adding a distinct chart for Nx Agents
As we narrow down our recommended deployment strategy for on-premise agents,
we are finding that the separate cluster is the more appropriate method
for reliability. This starts to break out the controller deployment and other
required elements from the core nx-cloud chart.

Once complete, we can publish a third nx-cloud-full chart that will contain
both other charts for the less optimal single-cluster solution.