chore(app): ensure most specific wins
I found that when trying to override a single image for the messagequeue
the global values continued to clobber mine as they are the last value
in the default clause. This moves the specific values to the end of the
default params to ensure if you have a more specific value it will
not be overridden by global values.