chore(nx-cloud): ensure custom volumes and mounts can be used
Internally we have a couple of extra mounts for things and we have to kustomize the yaml
output by helm. Instead let's just make it a bit easier by allowing users (and us) to
provide individual volumes. This is done in a way that still works with resource classes
and self signed certs.

Helpers were added as the conditionals were getting a bit long in the templates and tests
are included to ensure we still get what we expect.